+++ instrumentation/opentelemetry-instrumentation-botocore/src/opentelemetry/instrumentation/botocore/__init__.py
@@ -51,6 +51,7 @@ import logging
 from typing import Collection
 
 from botocore.client import BaseClient
+from botocore.endpoint import Endpoint
 from botocore.exceptions import ClientError
 from wrapt import wrap_function_wrapper
 
@@ -114,6 +115,7 @@ class BotocoreInstrumentor(BaseInstrumentor):
 
     def _uninstrument(self, **kwargs):
         unwrap(BaseClient, "_make_api_call")
+        unwrap(Endpoint, "prepare_request")
 
     @staticmethod
     def _is_lambda_invoke(service_name, operation_name, api_params):